Final commit
Well, the final one, already. It's been 12 weeks of development of this
student project, which were great and challenging. Too bad that we
couldn't have done everything as excepted.
Anyway, we have a good prototype and basis.

Elements from commit :
- Prefabs added
- Two test levels
- Material which fixes the display bug between tiles
- Added music
- Bug fixes

That was our game Project, Vendetta. Thanks for following the project.
- The :fr: guy.